Add test cases for academic years service

+++ academic_terms/src/main/java/com/sas/academic_terms/repository/AcademicYearRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AcademicYearRepository extends JpaRepository<AcademicYear, Integer> {
+public interface AcademicYearRepository extends JpaRepository<AcademicYear, Long> {
 }
